1.3.11 release

+++ modules/localgov_workflows_notifications/src/WorkflowNotification.php
@@ -42,10 +42,11 @@ class WorkflowNotification implements WorkflowNotificationInterface {
           continue;
         }
 
-        // Ensure the queue contains only one item for per service contact.
+        // Aggregate notifications by service contact and type.
         $found = FALSE;
+        $claimed_items = [];
         while ($queue_item = $queue->claimItem(1)) {
-          if ($queue_item->data->service_contact == $contact->id()) {
+          if ($queue_item->data->service_contact == $contact->id() && $queue_item->data->type === $type) {
 
             // Delete old item and create new one with additional entity.
             $queue->deleteItem($queue_item);
@@ -58,9 +59,15 @@ class WorkflowNotification implements WorkflowNotificationInterface {
             $found = TRUE;
             break;
           }
+          else {
+            $claimed_items[] = $queue_item;
+          }
         }
-        if ($queue_item !== FALSE && !$found) {
-          $queue->releaseItem($queue_item);
+
+        if ($claimed_items) {
+          foreach ($claimed_items as $queue_item) {
+            $queue->releaseItem($queue_item);
+          }
         }
 
         if (!$found) {
